@@ -10,25 +10,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.1.0",
-    "@angular/common": "^16.1.0",
-    "@angular/compiler": "^16.1.0",
-    "@angular/core": "^16.1.0",
-    "@angular/forms": "^16.1.0",
-    "@angular/platform-browser": "^16.1.0",
-    "@angular/platform-browser-dynamic": "^16.1.0",
-    "@angular/router": "^16.1.0",
+    "@angular/animations": "^17.1.3",
+    "@angular/common": "^17.1.3",
+    "@angular/compiler": "^17.1.3",
+    "@angular/core": "^17.1.3",
+    "@angular/forms": "^17.1.3",
+    "@angular/platform-browser": "^17.1.3",
+    "@angular/platform-browser-dynamic": "^17.1.3",
+    "@angular/router": "^17.1.3",
     "@iconify-json/octicon": "^1.1.52",
     "ngx-skeleton-loader": "^9.0.0",
     "octokit": "^3.1.1",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.8",
-    "@angular/cli": "~16.1.8",
-    "@angular/compiler-cli": "^16.1.0",
+    "@angular-devkit/build-angular": "^17.1.3",
+    "@angular/cli": "~17.1.3",
+    "@angular/compiler-cli": "^17.1.3",
     "@iconify/tailwind": "^0.1.4",
     "@types/jasmine": "~4.3.0",
     "autoprefixer": "^10.4.16",
