@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { ApiService } from 'src/app/services/api.service';
 
 @Component({
@@ -6,18 +6,17 @@ import { ApiService } from 'src/app/services/api.service';
   templateUrl: './profile.component.html',
   styleUrls: ['./profile.component.scss'],
 })
-export class ProfileComponent {
+export class ProfileComponent implements OnInit {
   // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-  user: any = null; // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-  repo: any = null;
-  constructor(private apiService: ApiService) {
+  user: any = null;
+  loader = true;
+
+  constructor(private apiService: ApiService) {}
+  ngOnInit(): void {
     // biome-ignore lint/suspicious/noExplicitAny: <explanation>
     this.apiService.getUser('johnpapa').subscribe((user: any) => {
       this.user = user;
-    });
-    // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-    this.apiService.getRepos('johnpapa').subscribe((repos: any) => {
-      this.repo = repos;
+      this.loader = false;
     });
   }
 }
